pyproject.toml
[project]
name = "easymem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.21.0",
    "fastembed>=0.6.1",
    "openai>=1.76.0",
    "qdrant-client[fastembed]>=1.12.1",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = [
    "ALL",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.6",
]
